tiny restructuring
