+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - Restructured the project directory layout
 - Moved from `setup.cfg` and `requirements.txt` to `poetry`
 - Added co-author 🙂
+- Changed the API to be closer to gymnasium
 
 ## [0.0.1] Initial Release
