+++ se_gym/__init__.py
@@ -1,11 +1,11 @@
-from .call_api import call_model
 from .observations import (
     ManualObserver,
     VectorStoreObserver,
     TreeSitterObeserver,
     CodeMapObserver,
 )
-from .caller import Sampler
-from .openai_lmu import get_openai_client
-from .executor import apply_patch, apply_patch_and_test, MalformedPatchException
+from .sampler import Sampler
+from .runner import apply_patch, apply_patch_and_test, MalformedPatchException
 from .fitness import percent_successfull, num_failed_tests
+from .api import make
+from . import openai_lmu
