+++ se_gym/config.py
@@ -1,0 +1,5 @@
+MAX_RETRIES = 5
+TIMEOUT_SECONDS = 30
+DEFAULT_SAVE_PATH = "./temp/"
+DOCKER_TAG = "pytest-env"
+GIT_APPLY_PATH_COMMAND = "git apply --ignore-space-change --ignore-whitespace --verbose --recount --inaccurate-eof ./file.patch"
